requirements.txt
# General requirements
Django>=5.0,<5.0.3
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.4,<2.9.7
pillow>=10.0.0,<10.3.0
